requirements.txt
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.26.4
opencv-python>=4.9.0
matplotlib>=3.8.4
einops>=0.8.0
tqdm>=4.66.4
tensorboard>=2.10.1
tensorboardX>=2.6.2
thop>=0.1.1
timm>=0.6.11
scikit-image>=0.19.1
Pillow>=10.3.0 